Reorganize queries in new file
